requirements.txt
python-telegram-bot==20.6
requests==2.31.0
matplotlib==3.8.0
python-dotenv==1.0.0
plotly==5.17.0
kaleido==0.2.1
pandas==2.1.1